+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enh #105: Raise the minimum PHP version to 8.0 (@xepozz, @rustamwin) 
 - Bug #107: Keep query string when generating from current route (@rustamwin)
+- Enh #108: Add `$queryParameters` parameter to `UrlGenerator::generateFromCurrent()` method (@rustamwin)
 
 ## 1.1.1 June 28, 2022